+++ eslint.config.js
@@ -3,4 +3,13 @@ const raycastConfig = require("@raycast/eslint-config");
 
 module.exports = defineConfig([
   ...raycastConfig,
+  {
+    ignores: ["raycast-env.d.ts"],
+  },
+  {
+    files: ["eslint.config.js"],
+    rules: {
+      "@typescript-eslint/no-require-imports": "off",
+    },
+  },
 ]);